Moved the reading of the delay into the effect specific code

+++ Effect.hpp
@@ -24,7 +24,7 @@ struct Effect
    void execute(void);
    void all_stop(uint8_t* buff);
    void reset();
-   uint16_t get_delay(const uint8_t lut[], size_t len);
+   uint8_t get_delay(const uint8_t lut[], size_t len);
 
    enum State
    {
